+++ vxGrid.h
@@ -76,7 +76,7 @@ protected:
 
 	vxVector3d m_position;
 	double m_size				= {5.0};
-	unsigned int m_resolution	= {5u};
+	unsigned long m_resolution	= {5u};
 	double m_invRes				= {1/5.0};
 
 	std::vector<vx>	m_data;
@@ -88,8 +88,8 @@ protected:
 	// cache objects
 	double m_boxSize			= {1.0};
 	double m_midBoxSize			= {.5};	
-	unsigned int m_resXres		= {25};
-	unsigned int m_resXresXres	= {125};
+	unsigned long m_resXres		= {25};
+	unsigned long m_resXresXres	= {125};
 	double m_xmin		= {0.0};
 	double m_xmax		= {0.0};
 	double m_ymin		= {0.0};
@@ -106,18 +106,18 @@ public:
 
 // GETTER AND SETTERS
 	
-	void createGridData(const unsigned int resolution);
-	void setResolution(unsigned int resolution);
+	void createGridData(const unsigned long resolution);
+	void setResolution(unsigned long resolution);
 	void setSize(const double size);
-	unsigned int size() const;
+	unsigned long size() const;
 	void setPosition(const vxVector3d position);
 	vxVector3d position() const;
-	unsigned int resolution() const;
+	unsigned long resolution() const;
 	void setBoxSize();
 	void updateBB();
 	void createDiagonals();
 	void createCorners();
-	void createGround(unsigned int offset = 0);
+	void createGround(unsigned long offset = 0);
 	void createEdges();
 	void createSphere(const vxVector3d &center, const double radio);
 	void createSphere(double x, double y, double z, const double radio);
@@ -128,64 +128,64 @@ public:
 					 const vxVector3d &scale);
 	
 // OPERATION WITH GRID
-	inline unsigned int index(const unsigned int x, 
-								const unsigned int y, 
-								const unsigned int z) const;
+	inline unsigned long index(const unsigned long x, 
+								const unsigned long y, 
+								const unsigned long z) const;
 	
 	static vxVector3d nextVoxel(const vxRay &ray, vxVector3d &exactCollision);
 	//sets every single vxl to 0.
 	void initialize(bool value = false);
 	//returns number of active voxels
-	unsigned int numActiveVoxels();
-	int getNumberOfVoxels() const;
+	unsigned long numActiveVoxels();
+	long getNumberOfVoxels() const;
 	// returns true if voxel at x y z is active
 	inline bool active(const vxVector3d& pos) const;
 	// returns true if voxel at index is active
-	inline bool active(unsigned int idx) const;
+	inline bool active(unsigned long idx) const;
 	// returns true if voxel at index is active
-	inline bool active(const unsigned int x, 
-					   const unsigned int y, 
-					   const unsigned int z) const;
+	inline bool active(const unsigned long x, 
+					   const unsigned long y, 
+					   const unsigned long z) const;
 	// sets active voxel at coordinates x y z
-	void activate(const unsigned int x, 
-				  const unsigned int y, 
-				  const unsigned int z);
+	void activate(const unsigned long x, 
+				  const unsigned long y, 
+				  const unsigned long z);
 	// sets active voxel at world space position
 	void activate(const vxVector3d &pos);
 	bool activeInRange(const vxVector3d &pos) const;
 
 	// sets unactive vxl at coordinates x y z
-	void deactivate(const unsigned int x, 
-					const unsigned int y, 
-					const unsigned int z);
+	void deactivate(const unsigned long x, 
+					const unsigned long y, 
+					const unsigned long z);
 
 	// returns true if element at local coords 
 	// is true
-	inline bool getElement(const unsigned int x, 
-						   const unsigned int y, 
-						   const unsigned int z) const;
+	inline bool getElement(const unsigned long x, 
+						   const unsigned long y, 
+						   const unsigned long z) const;
 	
-	void setElementColorIndex(const unsigned int x, const unsigned int y, const unsigned int z, const unsigned char c);
+	void setElementColorIndex(const unsigned long x, const unsigned long y, const unsigned long z, const unsigned char c);
 	// changes the value of the element at local
 	// coords x y z to be same as parameter value
-	inline void setElement(const unsigned int x, 
-						   const unsigned int y, 
-						   const unsigned int z, 
+	inline void setElement(const unsigned long x, 
+						   const unsigned long y, 
+						   const unsigned long z, 
 							bool value);
-	inline void setElement(unsigned int idx, bool value);
-	vxVector3d getVoxelPosition(const unsigned int iX, 
-								const unsigned int iY, 
-								const unsigned int iZ) const;
-	vxVector3d getVoxelPosition(unsigned int idx) const;
-	inline unsigned int indexAtPosition(const vxVector3d &position) const;
+	inline void setElement(unsigned long idx, bool value);
+	vxVector3d getVoxelPosition(const unsigned long iX, 
+								const unsigned long iY, 
+								const unsigned long iZ) const;
+	vxVector3d getVoxelPosition(unsigned long idx) const;
+	inline unsigned long indexAtPosition(const vxVector3d &position) const;
 	inline vx& vxAtPosition(const vxVector3d &position);
-	inline vx& vxAt(const unsigned int iX, 
-					const unsigned int iY, 
-					const unsigned int iZ);
+	inline vx& vxAt(const unsigned long iX, 
+					const unsigned long iY, 
+					const unsigned long iZ);
 	vx vxAtPosition(const vxVector3d &position) const;
-	vx vxAt(const unsigned int iX, 
-			const unsigned int iY, 
-			const unsigned int iZ) const;
+	vx vxAt(const unsigned long iX, 
+			const unsigned long iY, 
+			const unsigned long iZ) const;
 
 	bool inGrid(const vxVector3d &pnt, double tolerance) const;
 	bool inGrid(const vxVector3d &pnt) const;
@@ -194,7 +194,7 @@ public:
 	
 	int throwRay(const vxRay &ray, vxCollision &collide) const;
 	bool hasCollision(const vxRay &ray) const;
-	unsigned char elementColorIndex(const unsigned int x, const unsigned int y, const unsigned int z) const;
+	unsigned char elementColorIndex(const unsigned long x, const unsigned long y, const unsigned long z) const;
 };
 
 /*
